requirements.txt
streamlit
pandas
pandasql
numpy
scipy
scikit-learn
spacy>=3.0
textblob
python-docx
openpyxl
matplotlib
seaborn
lightgbm
networkx
shap
requests
sentence_transformers
torch
